Correct gegl_gtk_view_set_node’s transfer from full to none
The transfer of this function should be none because the helper’s
increases the node’s reference counter meaning that the caller still has
to unref it afterwards.

This caused a memory leak in the Vala example because
Gegl.Node.create_child returns an "unowned" node (i.e. the caller
doesn’t need to call unref) but Vala had to call ref on this node before
calling GeglGtk.View.set_node but didn’t call unref, as is the case when
passing an unowned variable to a full transfer function.  Thus, ref was
called twice (once by Vala and the other by the helper) but unref only
once (by the helper), creating the leak.

	*  gegl-gtk/gegl-gtk-view.c (gegl_gtk_view_set_node): Change
	   transfer annotation from full to none.